+++ plugins/behavior/nodegraph/src/Madgine/nodegraph/nodes/functionnode.h
@@ -15,15 +15,15 @@ namespace NodeGraph {
         std::string_view className() const override;
         std::unique_ptr<NodeBase> clone(NodeGraph &graph) const override;
 
-        size_t flowInCount(uint32_t group) const override;
+        uint32_t flowInCount(uint32_t group) const override;
 
-        size_t flowOutBaseCount(uint32_t group) const override;
+        uint32_t flowOutBaseCount(uint32_t group) const override;
 
-        virtual size_t dataInBaseCount(uint32_t group = 0) const override;
+        virtual uint32_t dataInBaseCount(uint32_t group = 0) const override;
         virtual std::string_view dataInName(uint32_t index, uint32_t group) const override;
         virtual ExtendedValueTypeDesc dataInType(uint32_t index, uint32_t group, bool bidir = true) const override;
 
-        virtual size_t dataProviderBaseCount(uint32_t group = 0) const override;
+        virtual uint32_t dataProviderBaseCount(uint32_t group = 0) const override;
         virtual ExtendedValueTypeDesc dataProviderType(uint32_t index, uint32_t group, bool bidir = true) const override;
 
         virtual void interpretRead(NodeInterpreterStateBase &interpreter, ValueType &retVal, std::unique_ptr<NodeInterpreterData> &data, uint32_t providerIndex, uint32_t group) const override;
