+++ plugins/behavior/nodegraph/src/Madgine/nodegraph/nodes/testnode.h
@@ -11,15 +11,15 @@ namespace NodeGraph {
         TestNode(NodeGraph &graph);
         TestNode(const TestNode &other, NodeGraph &graph);
 
-        virtual size_t flowInCount(uint32_t group) const override;
+        virtual uint32_t flowInCount(uint32_t group) const override;
 
-        virtual size_t flowOutBaseCount(uint32_t group) const override;
+        virtual uint32_t flowOutBaseCount(uint32_t group) const override;
 
-        virtual size_t dataInBaseCount(uint32_t group) const override;
+        virtual uint32_t dataInBaseCount(uint32_t group) const override;
         virtual std::string_view dataInName(uint32_t index, uint32_t group) const override;
         virtual ExtendedValueTypeDesc dataInType(uint32_t index, uint32_t group, bool bidir = true) const override;
 
-        virtual size_t dataOutBaseCount(uint32_t group) const override;
+        virtual uint32_t dataOutBaseCount(uint32_t group) const override;
         virtual std::string_view dataOutName(uint32_t index, uint32_t group) const override;
         virtual ExtendedValueTypeDesc dataOutType(uint32_t index, uint32_t group, bool bidir = true) const override;
